+++ src/jwks_cert__b.d.ts
@@ -1,3 +1,0 @@
-import type { B } from '@ctx-core/object'
-export declare const jwks_cert__b:B<jwks_cert__T>
-export type jwks_cert__T = ()=>Promise<string>