+++ src/jwks_x5c_.d.ts
@@ -1,0 +1,2 @@
+import type { Ctx } from '@ctx-core/object'
+export declare function jwks_x5c_(ctx:Ctx):Promise<string>