+++ src/jwks_x5c__b.d.ts
@@ -1,3 +1,0 @@
-import type { B } from '@ctx-core/object'
-export declare const jwks_x5c__b:B<jwks_x5c__T>
-export type jwks_x5c__T = ()=>Promise<string>