+++ src/jwt_token_decoded__b.d.ts
@@ -1,4 +1,0 @@
-import type { B } from '@ctx-core/object'
-import type { jwt_token_decoded_I } from '@ctx-core/auth0'
-export declare const jwt_token_decoded__b:B<jwt_token_decoded__T>
-export type jwt_token_decoded__T = (jwt_token:string)=>Promise<jwt_token_decoded_I>