+++ src/verify_jwt_email.d.ts
@@ -1,0 +1,2 @@
+import type { Ctx } from '@ctx-core/object'
+export declare function verify_jwt_email(ctx:Ctx, authorization:string):Promise<string>