+++ src/verify_jwt_email__b.d.ts
@@ -1,3 +1,0 @@
-import type { B } from '@ctx-core/object'
-export declare const verify_jwt_email__b:B<verify_jwt_email__T>
-export type verify_jwt_email__T = (authorization:string)=>Promise<string>