+++ src/verify_jwt_user_id__b.d.ts
@@ -1,3 +1,0 @@
-import { B } from '@ctx-core/object'
-export declare const verify_jwt_user_id__b:B<verify_jwt_user_id__T>
-export type verify_jwt_user_id__T = (authorization:string)=>Promise<string>